added one draft pseudo element builder – Locator.find as a more stable alternative to .find from Cypress
it's still an "instant" draft implementation, that will not be called in the chain, e.g.  after .by(selector),
because the original cypress' find will be called.

but already now, if called exactly after s(selector).find(selector).click(), it will ensure that inner selector resolves before performing the click.

under the hood it utilized a newly added condition to this commit: 'have.elements', that is simply JQuery's $.has(selector).length >= 0